+++ gtars/src/tokenizers/tokenizers/mod.rs
@@ -1,1 +1,18 @@
 pub mod tree_tokenizer;
+
+use std::path::{Path, PathBuf};
+
+pub trait Tokenizer {
+    /// Tokenize the given sequence into multiple underlying `Token`. The `offsets` on the `Token`
+    /// are expected to be relative to the given sequence.
+    fn tokenize(&self, sequence: &str) -> Result<Vec<TokenizedRegion>>;
+    /// Find the ID associated to a string token
+    fn token_to_id(&self, token: &str) -> Option<u32>;
+    /// Find the string token associated to an ID
+    fn id_to_token(&self, id: u32) -> Option<String>;
+    /// Retrieve the size of the vocabulary
+    fn get_vocab_size(&self) -> usize;
+    /// Save the current `Model` in the given folder, using the given `prefix` for the various
+    /// files that need to be saved.
+    fn save(&self, folder: &Path, prefix: Option<&str>) -> Result<Vec<PathBuf>>;
+}
